requirements.txt
llama-cpp-python
fastapi
uvicorn
pydantic
docstring_parser
pytest
pytest-asyncio
psutil
